[RUTA] Paso 3: Migrar blueprint expedientes a app/modules/ (#93 Fase 3)
- app/modules/expedientes/routes.py: Blueprint completo con template_folder='templates'
- Templates movidos (git mv): listado_v2, tramitacion_v3, detalle, editar
  de app/templates/expedientes/ a app/modules/expedientes/templates/expedientes/
- app/__init__.py: eliminado registro legacy de expedientes.bp desde app/routes/
- app/routes/expedientes.py: eliminado (migrado al módulo)

Los templates wizard_paso*.html permanecen en app/templates/expedientes/
ya que pertenecen al blueprint wizard_expediente (app/routes/).
